backend/requirements.txt
uvicorn==0.11.8
fastapi==0.61.0
pillow
numpy
python-multipart